+++ AquaEngine.API/Control/Interfaces/REST/Resources/UpdateProductOwnerResource.cs
@@ -1,3 +1,3 @@
 ﻿namespace AquaEngine.API.Control.Interfaces.REST.Resources;
 
-public record UpdateProductOwnerResource(long UserId, long ProductId);
+public record UpdateProductOwnerResource(int UserId, int ProductId);
